Move LobbyConstants to shared location 'lobby-client-data' (#14347)
Makes the class available to server.
