+++ game-app/domain-data/build.gradle.kts
@@ -1,3 +1,6 @@
 plugins {
     id("triplea-java-library")
 }
+dependencies {
+    implementation(project(":lobby-client-data"))
+}
