+++ game-app/domain-data/src/test/java/org/triplea/domain/data/UserNameTest.java
@@ -10,6 +10,7 @@ import java.util.List;
 import java.util.Optional;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
+import org.triplea.http.client.lobby.LobbyConstants;
 
 class UserNameTest {
