+++ game-app/game-headed/src/main/java/games/strategy/engine/lobby/client/login/ChangeEmailPanel.java
@@ -11,8 +11,8 @@ import javax.swing.JLabel;
 import javax.swing.JOptionPane;
 import javax.swing.JPanel;
 import javax.swing.JTextField;
-import org.triplea.domain.data.LobbyConstants;
 import org.triplea.domain.data.PlayerEmailValidation;
+import org.triplea.http.client.lobby.LobbyConstants;
 import org.triplea.http.client.web.socket.client.connections.PlayerToLobbyConnection;
 import org.triplea.swing.DialogBuilder;
 import org.triplea.swing.DocumentListenerBuilder;
