+++ game-app/game-headed/src/main/java/games/strategy/engine/lobby/client/login/ForgotPasswordPanel.java
@@ -16,9 +16,9 @@ import javax.swing.JLabel;
 import javax.swing.JOptionPane;
 import javax.swing.JPanel;
 import javax.swing.JTextField;
-import org.triplea.domain.data.LobbyConstants;
 import org.triplea.domain.data.PlayerEmailValidation;
 import org.triplea.domain.data.UserName;
+import org.triplea.http.client.lobby.LobbyConstants;
 import org.triplea.swing.JTextFieldBuilder;
 import org.triplea.swing.key.binding.KeyCode;
 import org.triplea.swing.key.binding.SwingKeyBinding;
