+++ game-app/game-headed/src/main/java/games/strategy/engine/lobby/client/login/LoginPanel.java
@@ -21,8 +21,8 @@ import javax.swing.JOptionPane;
 import javax.swing.JPanel;
 import javax.swing.JPasswordField;
 import javax.swing.JTextField;
-import org.triplea.domain.data.LobbyConstants;
 import org.triplea.domain.data.UserName;
+import org.triplea.http.client.lobby.LobbyConstants;
 import org.triplea.java.Sha512Hasher;
 import org.triplea.swing.JButtonBuilder;
 import org.triplea.swing.JCheckBoxBuilder;
